refactor: move LottoGame to domain package
